x_disp: fix pipeline not stalling when it should
The `dep` signal is generated so that it stalls the pipeline when:
- An instruction uses a certain register (rs1, rs2, or rs3).
- This register is marked on the scoreboard as unavailable (i.e., the
  core is waiting for the coprocessor to write its value).
- The coprocessor is not writing the value of the register in this very
  moment (i.e., the register won't be marked as available next cycle).

However, the 3rd condition isn't written properly, and it simply checks
the value of x_result_rd_i, without checking if it's valid
(x_result_valid_i = 1) and actually being written (x_result_we_i = 1).

This patch rewrites the condition to check scratchpad_d directly (i.e.,
the value scratchpad_q will have on the next clock cycle), which already
performs those two checks.  In other words, it checks whether the
register has been marked as "write pending" AND if it will still be
marked as "write pending" in the next clock cycle.

Fixes #10.